+++ src/lib/client/formEnhance.ts
@@ -1,10 +1,8 @@
 import { enhance, applyAction } from '$app/forms';
-import { afterNavigate, invalidateAll } from '$app/navigation';
+import { invalidateAll } from '$app/navigation';
 import { page } from '$app/stores';
 import type { ActionResult } from '@sveltejs/kit';
-import { isElementInViewport, scrollToAndCenter } from './elements.js';
 import { get, type Readable, type Writable } from 'svelte/store';
-import { onDestroy, tick } from 'svelte';
 import { browser } from '$app/environment';
 import {
   SuperFormError,
@@ -16,13 +14,8 @@ import { stringify } from 'devalue';
 import type { Entity } from '../schemaEntity.js';
 import type { FormOptions, SuperForm } from './index.js';
 import { clientValidation, validateField } from './clientValidation.js';
-
-enum FetchStatus {
-  Idle = 0,
-  Submitting = 1,
-  Delayed = 2,
-  Timeout = 3
-}
+import { Form } from './form.js';
+import { onDestroy } from 'svelte';
 
 export type FormUpdate = (
   result: Exclude<ActionResult, { type: 'error' }>,
@@ -39,6 +32,16 @@ export type SuperFormEventList<T extends AnyZodObject, M> = {
     SuperFormEvents<T, M>[Property]
   >[];
 };
+
+type ValidationResponse<
+  Success extends Record<string, unknown> | undefined = Record<
+    string,
+    // eslint-disable-next-line @typescript-eslint/no-explicit-any
+    any
+  >,
+  // eslint-disable-next-line @typescript-eslint/no-explicit-any
+  Invalid extends Record<string, unknown> | undefined = Record<string, any>
+> = { result: ActionResult<Success, Invalid> };
 
 export function cancelFlash<T extends AnyZodObject, M>(
   options: FormOptions<T, M>
@@ -124,199 +127,12 @@ export function formEnhance<T extends AnyZodObject, M>(
   }
   formEl.addEventListener('focusout', checkBlur);
 
-  const ErrorTextEvents = new Set<HTMLFormElement>();
-
-  function ErrorTextEvents_selectText(e: Event) {
-    const target = e.target as HTMLInputElement;
-    if (options.selectErrorText) target.select();
-  }
-
-  function ErrorTextEvents_addErrorTextListeners(formEl: HTMLFormElement) {
-    formEl.querySelectorAll('input').forEach((el) => {
-      el.addEventListener('invalid', ErrorTextEvents_selectText);
-    });
-  }
-
-  function ErrorTextEvents_removeErrorTextListeners(
-    formEl: HTMLFormElement
-  ) {
-    formEl
-      .querySelectorAll('input')
-      .forEach((el) =>
-        el.removeEventListener('invalid', ErrorTextEvents_selectText)
-      );
-  }
-
-  afterNavigate((nav) => {
-    if (nav.type == 'goto') {
-      htmlForm.completed(true);
-    }
-  });
+  const htmlForm = Form(formEl, { submitting, delayed, timeout }, options);
 
   onDestroy(() => {
-    ErrorTextEvents.forEach((formEl) =>
-      ErrorTextEvents_removeErrorTextListeners(formEl)
-    );
-    ErrorTextEvents.clear();
-
     formEl.removeEventListener('focusout', checkBlur);
-
-    htmlForm.completed(true);
   });
 
-  type ValidationResponse<
-    Success extends Record<string, unknown> | undefined = Record<
-      string,
-      // eslint-disable-next-line @typescript-eslint/no-explicit-any
-      any
-    >,
-    // eslint-disable-next-line @typescript-eslint/no-explicit-any
-    Invalid extends Record<string, unknown> | undefined = Record<string, any>
-  > = { result: ActionResult<Success, Invalid> };
-
-  /**
-   * @DCI-context
-   */
-  function Form(formEl: HTMLFormElement) {
-    function rebind() {
-      if (options.selectErrorText) {
-        const form = Form_element();
-        if (form && formEl !== form) {
-          ErrorTextEvents_removeErrorTextListeners(form);
-          ErrorTextEvents.delete(form);
-        }
-        if (!ErrorTextEvents.has(formEl)) {
-          ErrorTextEvents_addErrorTextListeners(formEl);
-          ErrorTextEvents.add(formEl);
-        }
-      }
-
-      Form = formEl;
-    }
-
-    let Form: {
-      querySelectorAll: (selector: string) => NodeListOf<HTMLElement>;
-      querySelector: (selector: string) => HTMLElement;
-      dataset: DOMStringMap;
-    };
-
-    function Form_element() {
-      return Form as HTMLFormElement;
-    }
-
-    function Form_shouldAutoFocus(userAgent: string) {
-      if (typeof options.autoFocusOnError === 'boolean')
-        return options.autoFocusOnError;
-      else return !/iPhone|iPad|iPod|Android/i.test(userAgent);
-    }
-
-    const Form_scrollToFirstError = async () => {
-      if (options.scrollToError == 'off') return;
-
-      const selector = options.errorSelector;
-      if (!selector) return;
-
-      // Wait for form to update with errors
-      await tick();
-
-      // Scroll to first form message, if not visible
-      let el: HTMLElement | null;
-      el = Form.querySelector(selector) as HTMLElement | null;
-      if (!el) return;
-      // Find underlying element if it is a FormGroup element
-      el = el.querySelector(selector) ?? el;
-
-      const nav = options.stickyNavbar
-        ? (document.querySelector(options.stickyNavbar) as HTMLElement)
-        : null;
-
-      if (typeof options.scrollToError != 'string') {
-        el.scrollIntoView(options.scrollToError);
-      } else if (!isElementInViewport(el, nav?.offsetHeight ?? 0)) {
-        scrollToAndCenter(el, undefined, options.scrollToError);
-      }
-
-      // Don't focus on the element if on mobile, it will open the keyboard
-      // and probably hide the error message.
-      if (!Form_shouldAutoFocus(navigator.userAgent)) return;
-
-      let focusEl;
-      focusEl = el;
-
-      if (
-        !['INPUT', 'SELECT', 'BUTTON', 'TEXTAREA'].includes(focusEl.tagName)
-      ) {
-        focusEl = focusEl.querySelector<HTMLElement>(
-          'input:not([type="hidden"]):not(.flatpickr-input), select, textarea'
-        );
-      }
-
-      if (focusEl) {
-        try {
-          focusEl.focus({ preventScroll: true });
-          if (options.selectErrorText && focusEl.tagName == 'INPUT') {
-            (focusEl as HTMLInputElement).select();
-          }
-        } catch (err) {
-          // Some hidden inputs like from flatpickr cannot be focused.
-        }
-      }
-    };
-
-    rebind();
-
-    {
-      let state: FetchStatus = FetchStatus.Idle;
-      let delayedTimeout: number, timeoutTimeout: number;
-
-      const setState = (s: typeof state) => {
-        state = s;
-        submitting.set(state >= FetchStatus.Submitting);
-        delayed.set(state >= FetchStatus.Delayed);
-        timeout.set(state >= FetchStatus.Timeout);
-      };
-
-      return {
-        submitting: () => {
-          rebind();
-          setState(
-            state != FetchStatus.Delayed
-              ? FetchStatus.Submitting
-              : FetchStatus.Delayed
-          );
-
-          // https://www.nngroup.com/articles/response-times-3-important-limits/
-          if (delayedTimeout) clearTimeout(delayedTimeout);
-          if (timeoutTimeout) clearTimeout(timeoutTimeout);
-
-          delayedTimeout = window.setTimeout(() => {
-            if (state == FetchStatus.Submitting)
-              setState(FetchStatus.Delayed);
-          }, options.delayMs);
-
-          timeoutTimeout = window.setTimeout(() => {
-            if (state == FetchStatus.Delayed) setState(FetchStatus.Timeout);
-          }, options.timeoutMs);
-        },
-
-        completed: (cancelled: boolean) => {
-          if (delayedTimeout) clearTimeout(delayedTimeout);
-          if (timeoutTimeout) clearTimeout(timeoutTimeout);
-          delayedTimeout = timeoutTimeout = 0;
-
-          setState(FetchStatus.Idle);
-          if (!cancelled) setTimeout(Form_scrollToFirstError);
-        },
-
-        scrollToFirstError: () => setTimeout(Form_scrollToFirstError),
-
-        isSubmitting: () =>
-          state === FetchStatus.Submitting || state === FetchStatus.Delayed
-      };
-    }
-  }
-
-  const htmlForm = Form(formEl);
   let currentRequest: AbortController | null;
 
   return enhance(formEl, async (submit) => {